+++ peer_web_frontend/src/app/team-list/panel/Sidebar.tsx
